Refactoring...

+++ wt_ip_adapter_anycast_address_amd64.go
@@ -8,8 +8,8 @@ package winipcfg
 // https://docs.microsoft.com/en-us/windows/desktop/api/iptypes/ns-iptypes-_ip_adapter_anycast_address_xp
 // Defined in iptypes.h
 type IP_ADAPTER_ANYCAST_ADDRESS_XP struct {
-	Length ULONG
-	Flags DWORD
+	Length uint32 // Windows type: ULONG
+	Flags uint32 // Windows type: DWORD
 	Next *IP_ADAPTER_ANYCAST_ADDRESS_XP
 	Address wtSocketAddress
 }
