+++ wt_ip_adapter_multicast_address_xp_amd64.go
@@ -8,8 +8,8 @@ package winipcfg
 // https://docs.microsoft.com/en-us/windows/desktop/api/iptypes/ns-iptypes-_ip_adapter_multicast_address_xp
 // IP_ADAPTER_MULTICAST_ADDRESS_XP defined in iptypes.h
 type wtIpAdapterMulticastAddressXp struct {
-	Length ULONG
-	Flags DWORD
+	Length uint32 // Windows type: ULONG
+	Flags uint32 // Windows type: DWORD
 	Next *wtIpAdapterMulticastAddressXp
 	Address wtSocketAddress
 }
