+++ wt_ip_adapter_wins_server_address_lh_amd64.go
@@ -8,8 +8,8 @@ package winipcfg
 // https://docs.microsoft.com/en-us/windows/desktop/api/iptypes/ns-iptypes-_ip_adapter_wins_server_address_lh
 // IP_ADAPTER_WINS_SERVER_ADDRESS_LH defined in iptypes.h
 type wtIpAdapterWinsServerAddressLh struct {
-	Length ULONG
-	Reserved DWORD
+	Length uint32 // Windows type: ULONG
+	Reserved uint32 // Windows type: DWORD
 	Next *wtIpAdapterWinsServerAddressLh
 	Address wtSocketAddress
 }
