+++ wt_mib_ipforward_table2_386.go
@@ -7,7 +7,7 @@ package winipcfg
 
 // https://docs.microsoft.com/en-us/windows/desktop/api/netioapi/ns-netioapi-_mib_ipforward_table2
 type MIB_IPFORWARD_TABLE2 struct {
-	NumEntries ULONG
+	NumEntries uint32 // Windows type: ULONG
 	// In 32-bit builds we have to artificially add an offset, in order to get the same size of the struct.
 	offset [4]byte
 	Table [ANY_SIZE]MIB_IPFORWARD_ROW2
