+++ wt_mib_ipforward_table2_amd64.go
@@ -7,6 +7,6 @@ package winipcfg
 
 // https://docs.microsoft.com/en-us/windows/desktop/api/netioapi/ns-netioapi-_mib_ipforward_table2
 type MIB_IPFORWARD_TABLE2 struct {
-	NumEntries ULONG
+	NumEntries uint32 // Windows type: ULONG
 	Table [ANY_SIZE]MIB_IPFORWARD_ROW2
 }
